Backend/requirements.txt
aniso8601==7.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.2.13
django-ckeditor==5.9.0
django-cors-headers==3.2.1
django-easy-select2==1.5.6
django-filter==2.2.0
django-graphql-jwt==0.3.0
django-graphql-social-auth==0.1.4
django-heroku==0.3.1
django-import-export==2.0.2
django-js-asset==1.2.2
et-xmlfile==1.0.1
googledrivedownloader==0.4
graphene==2.1.8
graphene-django==2.9.0
graphql-core==2.3.1
graphql-relay==2.0.1
gunicorn==20.0.4
idna==2.9
jdcal==1.4.1
MarkupPy==1.14
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.1
promise==2.3
psycopg2==2.8.4
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.3.1
tablib==1.1.0
urllib3==1.25.8
whitenoise==5.0.1
xlrd==1.2.0
xlwt==1.3.0
